+++ example/android/app/build.gradle
@@ -72,6 +72,10 @@ import com.android.build.OutputFile
  * ]
  */
 
+project.ext.react = [
+    entryFile: "index.js"
+]
+
 apply from: "../../node_modules/react-native/react.gradle"
 
 /**
@@ -90,13 +94,13 @@ def enableSeparateBuildPerCPUArchitecture = false
 def enableProguardInReleaseBuilds = false
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 23
+    buildToolsVersion "23.0.1"
 
     defaultConfig {
         applicationId "com.adjust.examples"
         minSdkVersion 16
-        targetSdkVersion 25
+        targetSdkVersion 22
         versionCode 1
         versionName "1.0"
         ndk {
@@ -135,10 +139,10 @@ android {
 dependencies {
     compile project(':react-native-adjust')
     compile fileTree(dir: "libs", include: ["*.jar"])
-    compile "com.android.support:appcompat-v7:25.0.1"
+    compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
-    compile 'com.google.android.gms:play-services-analytics:+'
-    compile 'com.android.installreferrer:installreferrer:+'
+    compile 'com.android.installreferrer:installreferrer:1.0'
+    compile 'com.google.android.gms:play-services-analytics:10.0.1'
 }
 
 // Run this once to be able to run the application with BUCK
